Add files via upload
This is a slightly edited version of a coding assignment for my c++ class at Macomb Community College. It is a program made for a surfboard shop that offers 4 sizes of surfboard, the program askes the user for information regarding the purchase then calculates their total cost and shows the individual cost of each size of the boards.